refactor model save and load methods
