--- .env
@@ -1,2 +1,0 @@
-MODEL_LOAD_FROM_FILE = "False"
-MODEL_PATH = "../resources/model"
