+++ lib/model/bert_truncated.py
@@ -1,8 +1,5 @@
 from __future__ import annotations
-import json
-from pathlib import Path
 from typing import Optional
-
 import torch
 from torch import Tensor
 from torch.nn import Module
@@ -17,10 +14,11 @@ class BertClassifierTruncated(BertClassifier):
         params: dict,
         tokenizer: Optional[AutoTokenizer] = None,
         neural_network: Optional[Module] = None,
+        pretrained_model_name_or_path: Optional[str] = "bert-base-uncased",
         device: str = "cuda:0",
         many_gpus: bool = False,
     ):
-        super().__init__(params, tokenizer, neural_network, device, many_gpus)
+        super().__init__(params, tokenizer, neural_network, pretrained_model_name_or_path, device, many_gpus)
 
     def _tokenize(self, texts: list[str]) -> BatchEncoding:
         """Transforms list of texts to list of tokens (truncated to 512 tokens)."""
@@ -37,22 +35,14 @@ class BertClassifierTruncated(BertClassifier):
         predictions = torch.flatten(predictions).cpu()
         return predictions
 
-    def save(self, model_dir: str) -> None:
-        model_dir = Path(model_dir)
-        model_dir.mkdir(parents=True, exist_ok=True)
-        with open(file=model_dir / "params.json", mode="w", encoding="utf-8") as file:
-            json.dump(self.params, file)
-        self.tokenizer.save_pretrained(model_dir)
-        if self.many_gpus:
-            torch.save(self.neural_network.module, model_dir / "model.bin")
-        else:
-            torch.save(self.neural_network, model_dir / "model.bin")
-
     @classmethod
     def load(cls, model_dir: str, device: str = "cuda:0", many_gpus: bool = False) -> BertClassifierTruncated:
-        model_dir = Path(model_dir)
-        with open(file=model_dir / "params.json", mode="r", encoding="utf-8") as file:
-            params = json.load(file)
-        tokenizer = AutoTokenizer.from_pretrained(model_dir)
-        neural_network = torch.load(f=model_dir / "model.bin", map_location=device)
-        return cls(params, tokenizer, neural_network, device, many_gpus)
+        model = super().load(model_dir, device, many_gpus)
+        return cls(
+            params=model.params,
+            tokenizer=model.tokenizer,
+            neural_network=model.neural_network,
+            pretrained_model_name_or_path=None,
+            device=model.device,
+            many_gpus=model.many_gpus,
+        )
