+++ requirements.txt
@@ -104,7 +104,6 @@ pyrsistent==0.19.3
 PySocks==1.7.1
 pytest==7.2.1
 python-dateutil==2.8.2
-python-dotenv==0.21.1
 python-json-logger==2.0.7
 pytz==2022.7.1
 PyYAML==6.0
